adversarial search agent to play the game "Isolation"
In this project, students will develop an adversarial search agent to
play the game "Isolation". Isolation is a deterministic, two-player game
of perfect information in which the players alternate turns moving a
single piece from one cell to another on a board. Whenever either player
occupies a cell, that cell becomes blocked for the remainder of the
game. The first player with no remaining legal moves loses, and the
opponent is declared the winner